// src/lib/pocketbase-types.ts
export enum Collections {
	Authorigins = "_authOrigins",
	Externalauths = "_externalAuths",
	Mfas = "_mfas",
	Otps = "_otps",
	Superusers = "_superusers",
	Cards = "cards",
	Decks = "decks",
	Guesses = "guesses",
	Tags = "tags",
	Users = "users",
}

// Alias types for improved usability
export type IsoDateString = string
export type RecordIdString = string
export type HTMLString = string

type ExpandType<T> = unknown extends T
	? T extends unknown
		? { expand?: unknown }
		: { expand: T }
	: { expand: T }

// System fields
export type BaseSystemFields<T = unknown> = {
	id: RecordIdString
	collectionId: string
	collectionName: Collections
} & ExpandType<T>
// ...
export type CardsRecord = {
	answer?: string
	created?: IsoDateString
	deckId?: RecordIdString
	id: string
	question?: string
	tags?: RecordIdString[]
	updated?: IsoDateString
}

export type DecksRecord = {
	created?: IsoDateString
	description?: string
	id: string
	public?: boolean
	title?: string
	updated?: IsoDateString
	userId?: RecordIdString
}

export type GuessesRecord = {
	cardId?: RecordIdString
	created?: IsoDateString
	id: string
	isCorrect?: boolean
	updated?: IsoDateString
	userId?: RecordIdString
}
// ...
export type CardsResponse<Texpand = unknown> = Required<CardsRecord> & BaseSystemFields<Texpand>
export type DecksResponse<Texpand = unknown> = Required<DecksRecord> & BaseSystemFields<Texpand>
export type GuessesResponse<Texpand = unknown> = Required<GuessesRecord> & BaseSystemFields<Texpand>